requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Pillow==12.2.0
